requirements.txt
azure-core==1.8.1
azure-storage-blob==12.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
filelock==3.0.12
future==0.18.2
idna==2.10
isodate==0.6.0
joblib==0.16.0
msrest==0.6.19
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
pycparser==2.20
pyparsing==2.4.7
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
sacremoses==0.0.43
sentencepiece==0.1.91
six==1.15.0
tokenizers==0.8.1rc2
torch==1.6.0
torchvision==0.7.0
tqdm==4.49.0
transformers==3.2.0
urllib3==1.25.10
